AC-333341: refactoring - split grouping login into an hierarchy of several groupers
